Refactored Fish class into part1/
This is due to the different implementation used in part 2.
