+++ day6/part1/fish.py
